Fix errors for servers not using CC plus names
Due to the fact you were using p.name here, this would suggest that all usernames have the ClassiCube + after it. If a player was found in the database to use a plus with a similar name as without, it did not compensate.

This commit really just replaces all p.name values with a locally-introduced isPlus variable which fixes all issues.

Tested with both + names on and off and no issues.